refact: move `KeysMapping` to `taf.models.types`
